Create README.md
#Unity Dependency Injection AKA UnityRnD
The purpose of this project is to test Unity DI when you have multiple projects and dll libraries. All of the examples on the web are boring calculators or couple all the dlls to the main application.

I am trying to show, in its simplest form, how to use Unity in your application when your application has no hard references to any outside libraries. It should also show how you can hot swap dlls or add dlls on the fly without rebuilding the main application.